Rename image_service to attribution_service
- Renamed app/services/image_service.py to app/services/attribution_service.py
- Updated import statement in app/api/routes.py
- Updated README.md documentation to reflect the new filename
- All functionality remains the same, just better naming consistency
