+++ app/services/attribution_service.py
